refactored tests, moved config file
