+++ package.json
@@ -4,7 +4,7 @@
   "description": "A miniscule MEAN stack app generated by Meaniscule",
   "main": "server.js",
   "scripts": {
-    "test": "mocha --harmony --recursive server/ tests/",
+    "test": "mocha --harmony --recursive tests/",
     "start": "nodemon --watch server server/start.js",
     "postinstall": "gulp build",
     "config-example": "node scripts/generate-config-example.js",
